Scout ciphers helper
Ok everybody.
Today I create this repository called my personal projects where I will put my great projects of programming.
So my first project is Scout cipher helper.
You know that scouts, soldiers, cops and etc uses many types of ciphers to talk with each other secretly.
This program is for scouts but you can use in other ways.
So now I will talk about how many ciphers I put in this program.
The first cipher is the homograph.
This is a cipher used to talk to a group of scouts at long distance.
The second cipher is the OTAN alphabet. This is used to talk in the radios normally is used to talk the name of the person that is using the walkie talkie but it can refer to groups like in the movies. eg: team Alpha over, team Romeo over.
The third cipher is the chinese cipher.
The vowels are vertical bars.
The consonants are horizontal bars.
In this cipher i tried to draw the chars but I did different. I used a custom font.
Don't worry. You don't need to install in your computer because the program reads the font automatically. I installed this font in the program.
The fourth cipher is the "false braille".
You know that braille is used to help blind people to read. But in the scouts you can use to encrypt messages.
This cipher works in this way:
First we divide the alphabet in three squares. In the program I divided in a tridimensional array.
We have in each square the column and the row.
The first ball is the square where is the letter.
The second ball is the row.
The third ball is the column
So in this project I was unable to find a font that contains this algorithm of encryption. So I created using GDI+(System.Drawing library)
We have the "DATE" cipher.
We digit a year and we organize the 26 letters of the alphabet by the numbers of the year.

The result will be for example: 112456
"11" is a char of the string.
And here I have the alphabet finally the more simple to implement algorithm.
The inverted alphabet.
I don't need explain what this means.
The cipher will be the inverted alphabet.
A will be Z.

Later I will implement more ciphers.