More ciphers
Now we have angular code.
I tried to create a font like braille code but I was unable because I don't know how to work with font editors.
The Braille code was more easy because FontStruct has prepared models to do our font. That's why I could create the false braille font.
Angular code works with perpendicular lines. The letters are divided by four squares like tic tac toe game the two first and the last two squares are made with the diagonals of the square. The second and the last square has a dot.
I did a angular keyboard to swap by the font.
In the activities it is more easy to decrypt. Each symbol is a letter.
They must insert the correspondent symbol and it will give the answer.
Bugs with the morse code and reversed alphabet fixed.