Transposed Alphabet
I add new ciphers to BPCipher.
I did a cipher called transposed alphabet.
You choose a letter and "A" is that letter.
There is a loop that goes from the letter to the end of the alphabet.
There is another loop that starts from 0 to the first letter index.
And this is the transposed alphabet.